[Viky] Add : Folder structure modified

+++ Example/LeapYear.cs
@@ -12,7 +12,7 @@ namespace Example
         {
             Console.WriteLine("Enter the year to check wheather is Leap or not ");
 
-            double year =Convert.ToDouble (Console.ReadLine());
+            int year =Convert.ToInt32 (Console.ReadLine());
 
             if ((year % 4 == 0 && year % 100 != 0 || year % 400 == 0))
             {
